ocr_service/requirements.txt
Pillow==9.3.0
pytesseract==0.3.1
pdf2image==1.11.0
numpy==1.22.0
tesserocr==2.7.0
fastapi==0.109.0
Image==1.5.33
pydantic==1.10.14
pyngrok==7.0.5
nltk==3.8.1
pypdfium2==4.21.0
uvicorn==0.21.0
python_docx==1.0.1
python-multipart